Runtime checks will only be performed once
It's not a good idea to perform this checks for every request.
Instead, they will be checked only once after booting up.
