+++ config/initializers/checks.rb
@@ -1,0 +1,2 @@
+# Clear the cache
+Rails.cache.write("portus-checks", nil)
